Moved go_aspect stuff to separate static/non-OO convenience module called go_utils
